reorganized akka-apps whiteboard, cleaned up caption and poll
