--- bbb-common-message/src/main/scala/org/bigbluebutton/common2/domain/Whiteboard.scala
@@ -1,3 +1,0 @@
-package org.bigbluebutton.common2.domain
-
-case class AnnotationVO(id: String, status: String, annotationType: String, annotationInfo: scala.collection.immutable.Map[String, Object], wbId: String, userId: String, position: Int)
