+++ bigbluebutton-apps/src/main/java/org/bigbluebutton/red5/pubsub/redis/RedisPubSubMessageHandler.java
@@ -3,7 +3,6 @@ package org.bigbluebutton.red5.pubsub.redis;
 import org.bigbluebutton.common.messages.MessagingConstants;
 import org.bigbluebutton.red5.client.MeetingClientMessageSender;
 import org.bigbluebutton.red5.client.UserClientMessageSender;
-import org.bigbluebutton.red5.client.WhiteboardClientMessageSender;
 import org.bigbluebutton.red5.client.messaging.IConnectionInvokerService;
 import org.bigbluebutton.red5.monitoring.BbbAppsIsKeepAliveHandler;
 import org.slf4j.Logger;
@@ -13,14 +12,12 @@ public class RedisPubSubMessageHandler implements MessageHandler {
 	private IConnectionInvokerService service;
 	private UserClientMessageSender userMessageSender;
 	private MeetingClientMessageSender meetingMessageSender;
-	private WhiteboardClientMessageSender whiteboardMessageSender;
 	//private BbbAppsIsKeepAliveHandler bbbAppsIsKeepAliveHandler;
 	
 	public void setConnectionInvokerService(IConnectionInvokerService s) {
 		this.service = s;
 		userMessageSender = new UserClientMessageSender(service);
 		meetingMessageSender = new MeetingClientMessageSender(service);
-		whiteboardMessageSender = new WhiteboardClientMessageSender(service);
 	}
 	
 	public void setBbbAppsIsKeepAliveHandler(BbbAppsIsKeepAliveHandler handler) {
@@ -34,8 +31,6 @@ public class RedisPubSubMessageHandler implements MessageHandler {
 			meetingMessageSender.handleMeetingMessage(message);
 		} else if (channel.equalsIgnoreCase(MessagingConstants.FROM_USERS_CHANNEL)) {
 			userMessageSender.handleUsersMessage(message);
-		} else if (channel.equalsIgnoreCase(MessagingConstants.FROM_WHITEBOARD_CHANNEL)) {
-			whiteboardMessageSender.handleWhiteboardMessage(message);
 		} else if (channel.equalsIgnoreCase(MessagingConstants.FROM_SYSTEM_CHANNEL)) {
 			//bbbAppsIsKeepAliveHandler.handleKeepAliveMessage(message);
 		}
